requirements.txt
flask==3.0.3
python-dateutil==2.9.0
peewee==3.17.7